+++ geo/CHANGES.md
@@ -16,6 +16,7 @@
 - Add `Bearing` and `Destination` trait implementations for `Euclidean`
 - Add `FillRule`-configurable boolean operations to `BooleanOps` trait
   - <https://github.com/georust/geo/pull/1382>
+- Fix panic in `algorithm::simplify::compute_rdp` with one point
 
 ##  Update `Intersections` with new implementation of the Bentley-Ottmann sweep-line algorithm to efficiently find sparse intersections between groups of lines.